webserver Update
Okay, so I think this code is right, but I have been spending waaay too much time hitting into a wall which I think is a memory limitation. Basically, I started working on making the code closer to functional, as well as trying to get to the point of being finished. Did not seem to add much.